refactor: split reducers into separate modules
