+++ src/components/card/Card.tsx
@@ -3,7 +3,7 @@ import 'react-loading-skeleton/dist/skeleton.css';
 import { Item } from '../../types/types';
 import './Card.css';
 import { DEFAULT_IMG } from '../../constants/constants';
-import { setIsPortalOpen } from '../../redux/reducers';
+import { setIsPortalOpen } from '../../redux/reducers/mainReducer';
 import { useAppDispatch } from '../../redux/hooks';
 
 type Props = {
