+++ src/components/card/Portal.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { DEFAULT_IMG, PORTAL_ERROR } from '../../constants/constants';
-import { setIsPortalOpen, useGetItemByIdQuery } from '../../redux/reducers';
+import { useGetItemByIdQuery } from '../../redux/reducers/api';
+import { setIsPortalOpen } from '../../redux/reducers/mainReducer';
 import Skeleton from 'react-loading-skeleton';
 import { useAppDispatch } from '../../redux/hooks';
