+++ src/components/searchBar/SearchBar.tsx
@@ -2,7 +2,7 @@ import React, { ChangeEvent, useEffect, useRef, useState } from 'react';
 import './search.css';
 import { useForm } from 'react-hook-form';
 import { useAppDispatch, useAppSelector } from '../../redux/hooks';
-import { setValue } from '../../redux/reducers';
+import { setValue } from '../../redux/reducers/mainReducer';
 
 const SearchBar = () => {
   const dispatch = useAppDispatch();
