+++ src/redux/types.ts
@@ -1,0 +1,5 @@
+import store from './store';
+
+export type RootState = ReturnType<typeof store.getState>;
+
+export type AppDispatch = typeof store.dispatch;
